reforge: split userSettings into modules
