+++ src/components/userProfileLayout/userProfileArchive/userProfileArchive.module.scss
@@ -1,0 +1,85 @@
+$mobileSize: 600px;
+
+.title {
+	font-size: 2.6rem;
+	line-height: 3.8rem;
+	color: var(--color-eerie-black);
+	align-self: flex-start;
+
+	@media (max-width: $mobileSize) {
+		font-size: 2rem;
+		line-height: 2.8rem;
+	}
+}
+
+.form {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 32px;
+
+	@media (max-width: 1440px) {
+		gap: 24px;
+	}
+
+	@media (max-width: $mobileSize) {
+		gap: 20px;
+	}
+}
+
+.form__settings {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 16px;
+	align-items: flex-start;
+	justify-content: flex-start;
+}
+
+.archive__items {
+	display: flex;
+	flex-direction: column;
+	gap: 14px;
+	width: 100%;
+	max-height: 400px;
+	overflow-y: auto;
+	overflow-x: hidden;
+
+	scrollbar-width: thin;
+	scrollbar-color: var(--color-light-green) var(--color-light-grayish-green);
+}
+
+.archiveItem {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: center;
+	width: 100%;
+	padding-right: 18px;
+}
+
+.archiveItem__title {
+	font-size: 1.6rem;
+	font-weight: 400;
+	line-height: 2.2rem;
+	color: var(--color-dark-cyan-blue);
+}
+
+.archiveItem__icons {
+	display: flex;
+	flex-direction: row;
+	gap: 16px;
+	align-items: center;
+}
+
+.archiveItem__icon {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	object-fit: contain;
+	cursor: pointer;
+}
+
+.resetWrapper {
+	position: relative;
+}
