+++ src/components/userProfileLayout/userProfileAvatarSettings/userProfileAvatarSettings.module.scss
@@ -1,0 +1,116 @@
+$mobileSize: 600px;
+
+.title {
+	font-size: 2.6rem;
+	line-height: 3.8rem;
+	color: var(--color-eerie-black);
+	align-self: flex-start;
+
+	@media (max-width: $mobileSize) {
+		font-size: 2rem;
+		line-height: 2.8rem;
+	}
+}
+
+.form {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 32px;
+
+	@media (max-width: 1440px) {
+		gap: 24px;
+	}
+
+	@media (max-width: $mobileSize) {
+		gap: 20px;
+	}
+}
+
+.form__settings {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 16px;
+	align-items: flex-start;
+	justify-content: flex-start;
+}
+
+.avatar__wrapper {
+	position: relative;
+	margin-bottom: 20px;
+	align-self: center;
+}
+
+.avatar__picture {
+	position: relative;
+}
+
+.avatar__image {
+	width: 110px;
+	height: 110px;
+}
+
+.avatar__editButtonWrapper {
+	position: absolute;
+	bottom: -16px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	width: 100%;
+}
+
+.avatar__editButton {
+	cursor: pointer;
+	width: 36px;
+	height: 36px;
+}
+
+.avatar__editIcon {
+	width: 36px;
+	height: 36px;
+}
+
+.avatar__input {
+	display: none;
+}
+
+.avatarTemplates {
+	align-self: center;
+	display: flex;
+	flex-direction: row;
+	flex-wrap: wrap;
+	column-gap: 30px;
+	row-gap: 12px;
+	align-content: center;
+	justify-content: center;
+	width: 100%;
+
+	@media (max-width: $mobileSize) {
+		gap: 20px;
+		max-width: 370px;
+	}
+}
+
+.templateAvatar__input {
+	display: none;
+}
+
+.templateAvatar__label {
+	display: flex;
+	cursor: pointer;
+
+	@media (max-width: $mobileSize) {
+		width: 60px;
+		height: 60px;
+	}
+}
+
+.avatarTemplates__img {
+	width: 80px;
+	height: 80px;
+	@media (max-width: $mobileSize) {
+		width: 60px;
+		height: 60px;
+	}
+}
