+++ src/components/userProfileLayout/userProfileChangePassword/userProfileChangePassword.module.scss
@@ -1,0 +1,37 @@
+$mobileSize: 600px;
+
+.title {
+	font-size: 2.6rem;
+	line-height: 3.8rem;
+	color: var(--color-eerie-black);
+	align-self: flex-start;
+
+	@media (max-width: $mobileSize) {
+		font-size: 2rem;
+		line-height: 2.8rem;
+	}
+}
+
+.form {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 32px;
+
+	@media (max-width: 1440px) {
+		gap: 24px;
+	}
+
+	@media (max-width: $mobileSize) {
+		gap: 20px;
+	}
+}
+
+.form__settings {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 16px;
+	align-items: flex-start;
+	justify-content: flex-start;
+}
