--- src/components/userProfileLayout/userProfileSettings/userProfileSettings.module.scss
@@ -1,264 +1,0 @@
-$mobileSize: 600px;
-
-.wrapper {
-	padding: 50px 40px;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-	height: 100%;
-
-	@media (max-width: 1440px) {
-		padding: 30px 40px;
-	}
-
-	@media (max-width: $mobileSize) {
-		padding: 16px;
-	}
-}
-
-.title {
-	font-size: 2.6rem;
-	line-height: 3.8rem;
-	color: var(--color-eerie-black);
-	align-self: flex-start;
-
-	@media (max-width: $mobileSize) {
-		font-size: 2rem;
-		line-height: 2.8rem;
-	}
-}
-
-.form {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 32px;
-
-	@media (max-width: 1440px) {
-		gap: 24px;
-	}
-
-	@media (max-width: $mobileSize) {
-		gap: 20px;
-	}
-}
-
-.form__settings {
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	gap: 16px;
-	align-items: flex-start;
-	justify-content: flex-start;
-}
-
-.saveButton {
-	margin-top: 16px;
-}
-
-.radioButtons {
-	display: flex;
-	flex-direction: row;
-	gap: 24px;
-	align-self: flex-start;
-}
-
-.userProfileNavigation {
-	display: flex;
-	flex-direction: column;
-	gap: 14px;
-}
-
-.navigationElementsWrap {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.navigationElements__title {
-	color: var(--color-dark-cyan-blue);
-	font-size: 20px;
-	font-weight: 400;
-}
-
-.userProfileFunctions {
-	display: flex;
-	flex-direction: column;
-	gap: 14px;
-}
-
-.profileFunctionsWrap {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	background: transparent;
-	border: none;
-}
-
-.profileFunctionsWrap__title {
-	font-size: 2rem;
-	line-height: 2.8rem;
-	font-weight: 400;
-	color: var(--color-dark-spring-green);
-	cursor: pointer;
-
-	@media (max-width: $mobileSize) {
-		font-size: 1.6rem;
-		line-height: 2.2rem;
-	}
-}
-
-.sidebarNav {
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	gap: 32px;
-}
-
-.archive__items {
-	display: flex;
-	flex-direction: column;
-	gap: 14px;
-	width: 100%;
-	max-height: 400px;
-	overflow-y: auto;
-	overflow-x: hidden;
-
-	scrollbar-width: thin;
-	scrollbar-color: var(--color-light-green) var(--color-light-grayish-green);
-}
-
-.archiveItem {
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	align-items: center;
-	width: 100%;
-	padding-right: 18px;
-}
-
-.archiveItem__title {
-	font-size: 1.6rem;
-	font-weight: 400;
-	line-height: 2.2rem;
-	color: var(--color-dark-cyan-blue);
-}
-
-.archiveItem__icons {
-	display: flex;
-	flex-direction: row;
-	gap: 16px;
-	align-items: center;
-}
-
-.archiveItem__icon {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	object-fit: contain;
-	cursor: pointer;
-}
-
-.resetWrapper {
-	position: relative;
-}
-
-.avatar__wrapper {
-	position: relative;
-	margin-bottom: 20px;
-	align-self: center;
-}
-
-.avatar__picture {
-	position: relative;
-}
-
-.avatar__image {
-	width: 110px;
-	height: 110px;
-}
-
-.avatar__editButtonWrapper {
-	position: absolute;
-	bottom: -16px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-}
-
-.avatar__editButton {
-	cursor: pointer;
-	width: 36px;
-	height: 36px;
-}
-
-.avatar__editIcon {
-	width: 36px;
-	height: 36px;
-}
-
-.avatar__input {
-	display: none;
-}
-
-.avatarTemplates {
-	align-self: center;
-	display: flex;
-	flex-direction: row;
-	flex-wrap: wrap;
-	column-gap: 30px;
-	row-gap: 12px;
-	align-content: center;
-	justify-content: center;
-	width: 100%;
-
-	@media (max-width: $mobileSize) {
-		gap: 20px;
-		max-width: 370px;
-	}
-}
-
-.templateAvatar__input {
-	display: none;
-}
-
-.templateAvatar__label {
-	display: flex;
-	cursor: pointer;
-
-	@media (max-width: $mobileSize) {
-		width: 60px;
-		height: 60px;
-	}
-}
-
-.avatarTemplates__img {
-	width: 80px;
-	height: 80px;
-	@media (max-width: $mobileSize) {
-		width: 60px;
-		height: 60px;
-	}
-}
-
-.removeButton {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	gap: 12px;
-	cursor: pointer;
-	margin-top: 10px;
-	margin-bottom: 10px;
-}
-
-.removeButton__title {
-	font-size: 2rem;
-	line-height: 2.8rem;
-	color: var(--color-bright-maroon);
-
-	@media (max-width: $mobileSize) {
-		font-size: 1.6rem;
-		line-height: 2.2rem;
-	}
-}
