+++ src/components/userProfileLayout/userProfileSidebar/userProfileSidebar.module.scss
@@ -145,6 +145,13 @@ $mobileSize: 600px;
 		}
 	}
 }
+
+.userInformationWrap {
+	@media (max-width: 1100px) {
+		cursor: pointer;
+	}
+}
+
 .userInformationWrap__date {
 	font-size: 2.2rem;
 	line-height: 3.5rem;
